+++ src/resources/documents.ts
@@ -1016,6 +1016,7 @@ export class DocumentResource extends BaseResource {
      * {
      *   "hash": "FE32EDDADE7CBDDCBB934E7402047450B0E59C02",
      *   "id": "63ddb172402799bfc991d10d",
+     *   "agreement_code": "550E8400-E29B-41D4-A716-446655440000", // printed on the certificate
      *   "status": "certificated",
      *   "page_count": "1",
      *   "signer_count": "1",